fullstack/backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
sqlalchemy>=1.4.0,<2.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
python-multipart>=0.0.9,<1.0.0
email-validator>=2.1.1,<3.0.0
